server/data-fetcher/requirements.txt
pandas
requests
SQLAlchemy
pymysql